Deduplicate StridedSlice Init and Prepare
The Init and Prepare functions were entirely common between the
reference kernel and the Xtensa optimized kernel. This PR relocates the
common functions to a strided_slice_common.cc file to be used on both
platforms. This reduces the reference kernel and xtena kernel files to
just the Eval function.

BUG=cleanup
